Seperated Login/Signup button on Navbar in its seperate component
