+++ components/Navbar/UserButton.tsx
@@ -1,0 +1,68 @@
+'use client'
+
+import { useState } from 'react'
+import LoginModal from 'components/Modals/Login'
+import SignUpModal from 'components/Modals/SignUp'
+import UserIcon from 'public/assets/icons/avatar.svg'
+import Avatar from 'components/Avatar'
+import { useUser, useHasMounted } from 'hooks'
+
+export default function UserButton() {
+  const hasMounted = useHasMounted()
+  const [openSignInModal, setOpenSignInModal] = useState(false)
+  const [openSignUpModal, setOpenSignUpModal] = useState(false)
+  const { user, isLoggedIn, isRegistered } = useUser()
+  const renderIconButton = () => {
+    if (!hasMounted) {
+      return <div className="ml-4 h-10 w-10" />
+    }
+
+    return isLoggedIn ? (
+      <button
+        onClick={() => setOpenSignInModal(true)}
+        className="text-grey-300 ml-4 h-10 w-10 cursor-pointer"
+      >
+        <Avatar avatar={user.avatar} size={30} />
+      </button>
+    ) : (
+      <button
+        onClick={() => toggleModal(true)}
+        className="text-grey-300 ml-4 h-10 w-10 cursor-pointer"
+      >
+        <UserIcon />
+      </button>
+    )
+  }
+
+  function toggleModal(show) {
+    if (user && user.publicKey && isRegistered) {
+      setOpenSignInModal(show)
+    } else {
+      setOpenSignUpModal(show)
+    }
+  }
+
+  function hideModals() {
+    setOpenSignInModal(false)
+    setOpenSignUpModal(false)
+  }
+
+  return (
+    <div>
+      <div>{renderIconButton()}</div>
+      {isRegistered ? (
+        <LoginModal
+          onLogin={() => setOpenSignInModal(false)}
+          open={openSignInModal}
+          onClose={hideModals}
+        />
+      ) : (
+        <SignUpModal
+          onConfirm={() => setOpenSignUpModal(false)}
+          open={openSignUpModal}
+          onClose={hideModals}
+        />
+      )}
+    </div>
+  )
+}
